3-Django-App/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
josepy==2.1.0
mozilla-django-oidc==4.0.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
